Fixed a crash resulting from deleting the buffers used for async loads before they are read from but after their existence if checked for.
Since there are two threads here, no matter how close I put the check for the existence of buffer to the read of the buffer, it still sometimes crashes because the load is cancelled and the buffer is deleted between the two events, meaning a non-existent buffer is read from.

A new object type has been created (oAsyncLoadBufferCleaner). When a chunk loader is deleted, instead of clearing its buffer, it makes this object and assigns that object's alarm 0 to be 1 frame LONGER than the longest possibe delay before an async load begins.

This means that when the load begins, the buffer will always be allowed to exist and if the async load event triggered on the same frame as deletion, then that's fine too because the buffer will still exist for several more frames.